Increase db client count
