+++ apps/api/src/lib/db.ts
@@ -1,0 +1,4 @@
+import { createDatabase, createPool } from "@echo-webkom/db/create";
+
+const pool = createPool();
+export const db = createDatabase(pool);
