+++ apps/api/src/services/feedback.ts
@@ -1,7 +1,6 @@
 import { Hono } from "hono";
 
-import { db } from "@echo-webkom/db";
-
+import { db } from "../lib/db";
 import { admin } from "../middleware/admin";
 
 const app = new Hono();
