+++ apps/web/src/actions/deregister.ts
@@ -3,8 +3,8 @@
 import { and, eq } from "drizzle-orm";
 import { z } from "zod";
 
-import { db } from "@echo-webkom/db";
 import { answers, registrations } from "@echo-webkom/db/schemas";
+import { db } from "@echo-webkom/db/serverless";
 import { DeregistrationNotificationEmail } from "@echo-webkom/email";
 import { emailClient } from "@echo-webkom/email/client";
