+++ apps/web/src/actions/whitelist.ts
@@ -2,8 +2,8 @@
 
 import { eq } from "drizzle-orm";
 
-import { db } from "@echo-webkom/db";
 import { insertWhitelistSchema, whitelist } from "@echo-webkom/db/schemas";
+import { db } from "@echo-webkom/db/serverless";
 
 import { getUser } from "@/lib/get-user";
 import { isMemberOf } from "@/lib/memberships";
