+++ apps/web/src/app/(default)/admin/whitelist/_actions/delete-access-request.ts
@@ -2,8 +2,8 @@
 
 import { eq } from "drizzle-orm";
 
-import { db } from "@echo-webkom/db";
 import { accessRequests } from "@echo-webkom/db/schemas";
+import { db } from "@echo-webkom/db/serverless";
 
 import { getUser } from "@/lib/get-user";
 import { isMemberOf } from "@/lib/memberships";
