+++ apps/web/src/app/(default)/auth/tilgang/[id]/_actions/request-access.ts
@@ -3,8 +3,9 @@
 import { nanoid } from "nanoid";
 import { z } from "zod";
 
-import { db, isPostgresIshError } from "@echo-webkom/db";
+import { isPostgresIshError } from "@echo-webkom/db/error";
 import { accessRequests } from "@echo-webkom/db/schemas";
+import { db } from "@echo-webkom/db/serverless";
 import { AccessRequestNotificationEmail } from "@echo-webkom/email";
 import { emailClient } from "@echo-webkom/email/client";
