+++ apps/web/src/data/access-requests/queries.ts
@@ -1,4 +1,4 @@
-import { db } from "@echo-webkom/db";
+import { db } from "@echo-webkom/db/serverless";
 
 export const getAccessRequests = async () => {
   return await db.query.accessRequests.findMany();
