+++ apps/web/src/data/reactions/queries.ts
@@ -1,7 +1,7 @@
 import { unstable_cache as cache } from "next/cache";
 import { eq } from "drizzle-orm";
 
-import { db } from "@echo-webkom/db";
+import { db } from "@echo-webkom/db/serverless";
 
 import { cacheKeyFactory } from "./revalidate";
