+++ packages/auth/src/helpers.ts
@@ -1,7 +1,7 @@
 import { eq } from "drizzle-orm";
 import { getServerSession as _getServerSession } from "next-auth/next";
 
-import { db } from "@echo-webkom/db";
+import { db } from "@echo-webkom/db/serverless";
 
 import { createAuthOptions } from "./auth-options";
