+++ packages/db/drizzle/migrate.ts
@@ -2,7 +2,7 @@
 import process from "node:process";
 import { migrate } from "drizzle-orm/postgres-js/migrator";
 
-import { db } from "../src";
+import { db } from "../src/serverless";
 
 /**
  * Do not run migrations in preview deployments.
