+++ packages/seeder/src/db/index.ts
@@ -1,7 +1,7 @@
 import chalk from "chalk";
 
-import { db } from "@echo-webkom/db";
 import { degrees, groups, spotRanges, usersToGroups } from "@echo-webkom/db/schemas";
+import { db } from "@echo-webkom/db/serverless";
 
 import * as message from "../utils";
 import { degrees as defaultDegrees } from "./data/degrees";
